requirements.txt
streamlit
pdfplumber
transformers
torch
gtts
Pillow
